+++ modern-treasury-java-core/src/main/kotlin/com/moderntreasury/api/models/InternalAccountUpdateParams.kt
@@ -32,6 +32,14 @@ private constructor(
     fun id(): Optional<String> = Optional.ofNullable(id)
 
     /**
+     * The Contra Ledger Account associated to this account.
+     *
+     * @throws ModernTreasuryInvalidDataException if the JSON field has an unexpected type (e.g. if
+     *   the server responded with an unexpected value).
+     */
+    fun contraLedgerAccountId(): Optional<String> = body.contraLedgerAccountId()
+
+    /**
      * The Counterparty associated to this account.
      *
      * @throws ModernTreasuryInvalidDataException if the JSON field has an unexpected type (e.g. if
@@ -71,6 +79,14 @@ private constructor(
      *   the server responded with an unexpected value).
      */
     fun parentAccountId(): Optional<String> = body.parentAccountId()
+
+    /**
+     * Returns the raw JSON value of [contraLedgerAccountId].
+     *
+     * Unlike [contraLedgerAccountId], this method doesn't throw if the JSON field has an unexpected
+     * type.
+     */
+    fun _contraLedgerAccountId(): JsonField<String> = body._contraLedgerAccountId()
 
     /**
      * Returns the raw JSON value of [counterpartyId].
@@ -154,14 +170,30 @@ private constructor(
          *
          * This is generally only useful if you are already constructing the body separately.
          * Otherwise, it's more convenient to use the top-level setters instead:
+         * - [contraLedgerAccountId]
          * - [counterpartyId]
          * - [ledgerAccountId]
          * - [metadata]
          * - [name]
-         * - [parentAccountId]
          * - etc.
          */
         fun body(body: InternalAccountUpdateRequest) = apply { this.body = body.toBuilder() }
+
+        /** The Contra Ledger Account associated to this account. */
+        fun contraLedgerAccountId(contraLedgerAccountId: String) = apply {
+            body.contraLedgerAccountId(contraLedgerAccountId)
+        }
+
+        /**
+         * Sets [Builder.contraLedgerAccountId] to an arbitrary JSON value.
+         *
+         * You should usually call [Builder.contraLedgerAccountId] with a well-typed [String] value
+         * instead. This method is primarily for setting the field to an undocumented or not yet
+         * supported value.
+         */
+        fun contraLedgerAccountId(contraLedgerAccountId: JsonField<String>) = apply {
+            body.contraLedgerAccountId(contraLedgerAccountId)
+        }
 
         /** The Counterparty associated to this account. */
         fun counterpartyId(counterpartyId: String) = apply { body.counterpartyId(counterpartyId) }
@@ -381,6 +413,7 @@ private constructor(
     class InternalAccountUpdateRequest
     @JsonCreator(mode = JsonCreator.Mode.DISABLED)
     private constructor(
+        private val contraLedgerAccountId: JsonField<String>,
         private val counterpartyId: JsonField<String>,
         private val ledgerAccountId: JsonField<String>,
         private val metadata: JsonField<Metadata>,
@@ -391,6 +424,9 @@ private constructor(
 
         @JsonCreator
         private constructor(
+            @JsonProperty("contra_ledger_account_id")
+            @ExcludeMissing
+            contraLedgerAccountId: JsonField<String> = JsonMissing.of(),
             @JsonProperty("counterparty_id")
             @ExcludeMissing
             counterpartyId: JsonField<String> = JsonMissing.of(),
@@ -404,7 +440,24 @@ private constructor(
             @JsonProperty("parent_account_id")
             @ExcludeMissing
             parentAccountId: JsonField<String> = JsonMissing.of(),
-        ) : this(counterpartyId, ledgerAccountId, metadata, name, parentAccountId, mutableMapOf())
+        ) : this(
+            contraLedgerAccountId,
+            counterpartyId,
+            ledgerAccountId,
+            metadata,
+            name,
+            parentAccountId,
+            mutableMapOf(),
+        )
+
+        /**
+         * The Contra Ledger Account associated to this account.
+         *
+         * @throws ModernTreasuryInvalidDataException if the JSON field has an unexpected type (e.g.
+         *   if the server responded with an unexpected value).
+         */
+        fun contraLedgerAccountId(): Optional<String> =
+            contraLedgerAccountId.getOptional("contra_ledger_account_id")
 
         /**
          * The Counterparty associated to this account.
@@ -446,6 +499,16 @@ private constructor(
          *   if the server responded with an unexpected value).
          */
         fun parentAccountId(): Optional<String> = parentAccountId.getOptional("parent_account_id")
+
+        /**
+         * Returns the raw JSON value of [contraLedgerAccountId].
+         *
+         * Unlike [contraLedgerAccountId], this method doesn't throw if the JSON field has an
+         * unexpected type.
+         */
+        @JsonProperty("contra_ledger_account_id")
+        @ExcludeMissing
+        fun _contraLedgerAccountId(): JsonField<String> = contraLedgerAccountId
 
         /**
          * Returns the raw JSON value of [counterpartyId].
@@ -515,6 +578,7 @@ private constructor(
         /** A builder for [InternalAccountUpdateRequest]. */
         class Builder internal constructor() {
 
+            private var contraLedgerAccountId: JsonField<String> = JsonMissing.of()
             private var counterpartyId: JsonField<String> = JsonMissing.of()
             private var ledgerAccountId: JsonField<String> = JsonMissing.of()
             private var metadata: JsonField<Metadata> = JsonMissing.of()
@@ -524,6 +588,7 @@ private constructor(
 
             @JvmSynthetic
             internal fun from(internalAccountUpdateRequest: InternalAccountUpdateRequest) = apply {
+                contraLedgerAccountId = internalAccountUpdateRequest.contraLedgerAccountId
                 counterpartyId = internalAccountUpdateRequest.counterpartyId
                 ledgerAccountId = internalAccountUpdateRequest.ledgerAccountId
                 metadata = internalAccountUpdateRequest.metadata
@@ -531,6 +596,21 @@ private constructor(
                 parentAccountId = internalAccountUpdateRequest.parentAccountId
                 additionalProperties =
                     internalAccountUpdateRequest.additionalProperties.toMutableMap()
+            }
+
+            /** The Contra Ledger Account associated to this account. */
+            fun contraLedgerAccountId(contraLedgerAccountId: String) =
+                contraLedgerAccountId(JsonField.of(contraLedgerAccountId))
+
+            /**
+             * Sets [Builder.contraLedgerAccountId] to an arbitrary JSON value.
+             *
+             * You should usually call [Builder.contraLedgerAccountId] with a well-typed [String]
+             * value instead. This method is primarily for setting the field to an undocumented or
+             * not yet supported value.
+             */
+            fun contraLedgerAccountId(contraLedgerAccountId: JsonField<String>) = apply {
+                this.contraLedgerAccountId = contraLedgerAccountId
             }
 
             /** The Counterparty associated to this account. */
@@ -631,6 +711,7 @@ private constructor(
              */
             fun build(): InternalAccountUpdateRequest =
                 InternalAccountUpdateRequest(
+                    contraLedgerAccountId,
                     counterpartyId,
                     ledgerAccountId,
                     metadata,
@@ -647,6 +728,7 @@ private constructor(
                 return@apply
             }
 
+            contraLedgerAccountId()
             counterpartyId()
             ledgerAccountId()
             metadata().ifPresent { it.validate() }
@@ -671,7 +753,8 @@ private constructor(
          */
         @JvmSynthetic
         internal fun validity(): Int =
-            (if (counterpartyId.asKnown().isPresent) 1 else 0) +
+            (if (contraLedgerAccountId.asKnown().isPresent) 1 else 0) +
+                (if (counterpartyId.asKnown().isPresent) 1 else 0) +
                 (if (ledgerAccountId.asKnown().isPresent) 1 else 0) +
                 (metadata.asKnown().getOrNull()?.validity() ?: 0) +
                 (if (name.asKnown().isPresent) 1 else 0) +
@@ -683,6 +766,7 @@ private constructor(
             }
 
             return other is InternalAccountUpdateRequest &&
+                contraLedgerAccountId == other.contraLedgerAccountId &&
                 counterpartyId == other.counterpartyId &&
                 ledgerAccountId == other.ledgerAccountId &&
                 metadata == other.metadata &&
@@ -693,6 +777,7 @@ private constructor(
 
         private val hashCode: Int by lazy {
             Objects.hash(
+                contraLedgerAccountId,
                 counterpartyId,
                 ledgerAccountId,
                 metadata,
@@ -705,7 +790,7 @@ private constructor(
         override fun hashCode(): Int = hashCode
 
         override fun toString() =
-            "InternalAccountUpdateRequest{counterpartyId=$counterpartyId, ledgerAccountId=$ledgerAccountId, metadata=$metadata, name=$name, parentAccountId=$parentAccountId, additionalProperties=$additionalProperties}"
+            "InternalAccountUpdateRequest{contraLedgerAccountId=$contraLedgerAccountId, counterpartyId=$counterpartyId, ledgerAccountId=$ledgerAccountId, metadata=$metadata, name=$name, parentAccountId=$parentAccountId, additionalProperties=$additionalProperties}"
     }
 
     /**