+++ modern-treasury-java-core/src/test/kotlin/com/moderntreasury/api/models/ExpectedPaymentCreateParamsTest.kt
@@ -15,6 +15,12 @@ internal class ExpectedPaymentCreateParamsTest {
     fun create() {
         ExpectedPaymentCreateParams.builder()
             .amountLowerBound(0L)
+            .amountReconciled(0L)
+            .amountReconciledDirection(ExpectedPaymentCreateParams.AmountReconciledDirection.CREDIT)
+            .amountUnreconciled(0L)
+            .amountUnreconciledDirection(
+                ExpectedPaymentCreateParams.AmountUnreconciledDirection.CREDIT
+            )
             .amountUpperBound(0L)
             .counterpartyId("182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e")
             .currency(Currency.AED)
@@ -126,6 +132,14 @@ internal class ExpectedPaymentCreateParamsTest {
         val params =
             ExpectedPaymentCreateParams.builder()
                 .amountLowerBound(0L)
+                .amountReconciled(0L)
+                .amountReconciledDirection(
+                    ExpectedPaymentCreateParams.AmountReconciledDirection.CREDIT
+                )
+                .amountUnreconciled(0L)
+                .amountUnreconciledDirection(
+                    ExpectedPaymentCreateParams.AmountUnreconciledDirection.CREDIT
+                )
                 .amountUpperBound(0L)
                 .counterpartyId("182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e")
                 .currency(Currency.AED)
@@ -236,6 +250,12 @@ internal class ExpectedPaymentCreateParamsTest {
         val body = params._body()
 
         assertThat(body.amountLowerBound()).contains(0L)
+        assertThat(body.amountReconciled()).contains(0L)
+        assertThat(body.amountReconciledDirection())
+            .contains(ExpectedPaymentCreateParams.AmountReconciledDirection.CREDIT)
+        assertThat(body.amountUnreconciled()).contains(0L)
+        assertThat(body.amountUnreconciledDirection())
+            .contains(ExpectedPaymentCreateParams.AmountUnreconciledDirection.CREDIT)
         assertThat(body.amountUpperBound()).contains(0L)
         assertThat(body.counterpartyId()).contains("182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e")
         assertThat(body.currency()).contains(Currency.AED)