+++ modern-treasury-java-core/src/test/kotlin/com/moderntreasury/api/services/blocking/InternalAccountServiceTest.kt
@@ -104,6 +104,7 @@ internal class InternalAccountServiceTest {
             internalAccountService.update(
                 InternalAccountUpdateParams.builder()
                     .id("id")
+                    .contraLedgerAccountId("contra_ledger_account_id")
                     .counterpartyId("counterparty_id")
                     .ledgerAccountId("ledger_account_id")
                     .metadata(
@@ -132,6 +133,21 @@ internal class InternalAccountServiceTest {
         val page = internalAccountService.list()
 
         page.items().forEach { it.validate() }
+    }
+
+    @Test
+    fun requestClosure() {
+        val client =
+            ModernTreasuryOkHttpClient.builder()
+                .baseUrl(TestServerExtension.BASE_URL)
+                .apiKey("My API Key")
+                .organizationId("my-organization-ID")
+                .build()
+        val internalAccountService = client.internalAccounts()
+
+        val internalAccount = internalAccountService.requestClosure("id")
+
+        internalAccount.validate()
     }
 
     @Test